vkeyboard: package for virtual keyboard
The vkeyboard package defines a uinput.Keyboard to use for key inputs.

Unlike the gousb package, the uinput wrapper [1] used here has been
archived as of May 2024.  The only other alternative [2] hasn't had any
updates since 2020.  This doesn't seem to be an issue right now, but
it's something to keep in mind for the future.

[1] https://github.com/bendahl/uinput
[2] https://github.com/sashko/go-uinput