device: support setting LCD image
Add methods for defining the image on the LCD.  The G13Device.SetLCD()
method takes an image.Image from the standard library and converts it to
the layout required for the G13.

The layout of the data in relation to the display layout is a bit
unusual.  The comment in imageToG13Bytes() explains the layout and how
it works.