config: handle stick when getting key states
If the stick is in Keys mode, return any keys that are mapped to the
position of the stick.
Keys can be assigned to the four directions of the stick: up, down,
left, and right.  The stick position ranges from 0 to 255 on each axis.
A direction is considered active if it is within 64 of the maximum value
in each direction.  This makes it easier to put the stick in a diagonal
position, activating two keys.
The active zone will probably become configurable in the future.